requirements.txt
onnxruntime
numpy
torchvision
Pillow
ijson